e-commerce-server/pyproject.toml
[project]
name = "e-commerce-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.0",
    "passlib",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.13.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.44.0",
]
